Move nowPlayingInfoPlaybackPublisher
